Lots of new abilities and status effects party recruitment lots of little bug fixes

+++ Models/Engine/Combat/StatusEffect.cs
@@ -9,13 +9,13 @@ namespace MerchantRPG.Models.Engine.Combat
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public int DamageModifier { get; set; }
+        public double DamageModifier { get; set; }
         public Action<Character> EffectAction { get; set; } = x => { };
         public int Turns { get; set; }
         public double MissPercent;
 
 
-        public StatusEffect(string _name, string _description, int _dex = 0, int _str = 0, int _int = 0, int _dmgMod = 1, int _turns = -1, double _missPercent = 0)
+        public StatusEffect(string _name, string _description, int _dex = 0, int _str = 0, int _int = 0, double _dmgMod = 1, int _turns = -1, double _missPercent = 0)
         {
             Name = _name;
             Description = _description;
